upgrade(installer): Add state transition commands

+++ pkg/fleet/installer/installer.go
@@ -26,6 +26,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/db"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 	installerErrors "github.com/DataDog/datadog-agent/pkg/fleet/installer/errors"
+	iexec "github.com/DataDog/datadog-agent/pkg/fleet/installer/exec"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/oci"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/datadogagent"
@@ -70,8 +71,14 @@ type Installer interface {
 	InstrumentAPMInjector(ctx context.Context, method string) error
 	UninstrumentAPMInjector(ctx context.Context, method string) error
 
-	Postinst(ctx context.Context, pkg string, caller string) error
-	Prerm(ctx context.Context, pkg string, caller string, update bool) error
+	PostInstall(ctx context.Context, pkg string, caller string) error
+	PreRemove(ctx context.Context, pkg string, caller string, update bool) error
+	PreStartExperiment(ctx context.Context, pkg string) error
+	PostStartExperiment(ctx context.Context, pkg string) error
+	PreStopExperiment(ctx context.Context, pkg string) error
+	PostStopExperiment(ctx context.Context, pkg string) error
+	PrePromoteExperiment(ctx context.Context, pkg string) error
+	PostPromoteExperiment(ctx context.Context, pkg string) error
 
 	Close() error
 }
@@ -90,15 +97,23 @@ type installerImpl struct {
 	userConfigsDir string
 }
 
+func newDB() (*db.PackagesDB, error) {
+	db, err := db.New(filepath.Join(paths.PackagesPath, "packages.db"), db.WithTimeout(10*time.Second))
+	if err != nil {
+		return nil, fmt.Errorf("could not create packages db: %w", err)
+	}
+	return db, nil
+}
+
 // NewInstaller returns a new Package Manager.
 func NewInstaller(env *env.Env) (Installer, error) {
 	err := ensureRepositoriesExist()
 	if err != nil {
 		return nil, fmt.Errorf("could not ensure packages and config directory exists: %w", err)
 	}
-	db, err := db.New(filepath.Join(paths.PackagesPath, "packages.db"), db.WithTimeout(10*time.Second))
+	db, err := newDB()
 	if err != nil {
-		return nil, fmt.Errorf("could not create packages db: %w", err)
+		return nil, err
 	}
 	i := &installerImpl{
 		env:        env,
@@ -370,6 +385,11 @@ func (i *installerImpl) InstallExperiment(ctx context.Context, url string) error
 			fmt.Errorf("could not extract package config layer: %w", err),
 		)
 	}
+
+	if err := i.preStartExperiment(ctx, pkg.Name); err != nil {
+		return fmt.Errorf("could not install experiment: %w", err)
+	}
+
 	repository := i.packages.Get(pkg.Name)
 	err = repository.SetExperiment(ctx, pkg.Version, tmpDir)
 	if err != nil {
@@ -379,7 +399,7 @@ func (i *installerImpl) InstallExperiment(ctx context.Context, url string) error
 		)
 	}
 
-	return i.startExperiment(ctx, pkg.Name)
+	return i.postStartExperiment(ctx, pkg.Name)
 }
 
 // RemoveExperiment removes an experiment.
@@ -388,8 +408,17 @@ func (i *installerImpl) RemoveExperiment(ctx context.Context, pkg string) error 
 	defer i.m.Unlock()
 
 	repository := i.packages.Get(pkg)
+	state, err := repository.GetState()
+	if err != nil {
+		return fmt.Errorf("could not get repository state: %w", err)
+	}
+	if !state.HasExperiment() {
+		// Fail early
+		return fmt.Errorf("no experiment to promote")
+	}
+
 	if runtime.GOOS != "windows" && pkg == packageDatadogInstaller {
-		// Special case for the Linux installer since `stopExperiment`
+		// Special case for the Linux installer since `preStopExperiment`
 		// will kill the current process, delete the experiment first.
 		err := repository.DeleteExperiment(ctx)
 		if err != nil {
@@ -398,12 +427,12 @@ func (i *installerImpl) RemoveExperiment(ctx context.Context, pkg string) error 
 				fmt.Errorf("could not delete experiment: %w", err),
 			)
 		}
-		err = i.stopExperiment(ctx, pkg)
+		err = i.preStopExperiment(ctx, pkg)
 		if err != nil {
 			return fmt.Errorf("could not stop experiment: %w", err)
 		}
 	} else {
-		err := i.stopExperiment(ctx, pkg)
+		err := i.preStopExperiment(ctx, pkg)
 		if err != nil {
 			return fmt.Errorf("could not stop experiment: %w", err)
 		}
@@ -415,7 +444,7 @@ func (i *installerImpl) RemoveExperiment(ctx context.Context, pkg string) error 
 			)
 		}
 	}
-	return nil
+	return i.postStopExperiment(ctx, pkg)
 }
 
 // PromoteExperiment promotes an experiment to stable.
@@ -424,17 +453,32 @@ func (i *installerImpl) PromoteExperiment(ctx context.Context, pkg string) error
 	defer i.m.Unlock()
 
 	repository := i.packages.Get(pkg)
-	err := repository.PromoteExperiment(ctx)
+	state, err := repository.GetState()
+	if err != nil {
+		return fmt.Errorf("could not get repository state: %w", err)
+	}
+	if !state.HasExperiment() {
+		// Fail early
+		return fmt.Errorf("no experiment to promote")
+	}
+
+	err = i.prePromoteExperiment(ctx, pkg)
 	if err != nil {
 		return fmt.Errorf("could not promote experiment: %w", err)
 	}
-	err = i.promoteExperiment(ctx, pkg)
+
+	err = repository.PromoteExperiment(ctx)
+	if err != nil {
+		return fmt.Errorf("could not promote experiment: %w", err)
+	}
+
+	err = i.postPromoteExperiment(ctx, pkg)
 	if err != nil {
 		return err
 	}
 
 	// Update db
-	state, err := repository.GetState()
+	state, err = repository.GetState()
 	if err != nil {
 		return fmt.Errorf("could not get repository state: %w", err)
 	}
@@ -480,7 +524,7 @@ func (i *installerImpl) InstallConfigExperiment(ctx context.Context, pkg string,
 	case "windows":
 		return nil // TODO: start config experiment for Windows
 	default:
-		return i.startExperiment(ctx, pkg)
+		return i.postStartExperiment(ctx, pkg)
 	}
 }
 
@@ -489,7 +533,7 @@ func (i *installerImpl) RemoveConfigExperiment(ctx context.Context, pkg string) 
 	i.m.Lock()
 	defer i.m.Unlock()
 
-	err := i.stopExperiment(ctx, pkg)
+	err := i.preStopExperiment(ctx, pkg)
 	if err != nil {
 		return fmt.Errorf("could not stop experiment: %w", err)
 	}
@@ -521,7 +565,7 @@ func (i *installerImpl) PromoteConfigExperiment(ctx context.Context, pkg string)
 	if err != nil {
 		log.Warnf("could not write user-facing config symlinks: %v", err)
 	}
-	return i.promoteExperiment(ctx, pkg)
+	return i.postPromoteExperiment(ctx, pkg)
 }
 
 // Purge removes all packages.
@@ -688,8 +732,8 @@ func (i *installerImpl) close() error {
 	return nil
 }
 
-// Postinst runs the post-install script for a package.
-func (i *installerImpl) Postinst(ctx context.Context, pkg string, caller string) error {
+// PostInstall runs the post-install command for a package.
+func (i *installerImpl) PostInstall(ctx context.Context, pkg string, caller string) error {
 	i.m.Lock()
 	defer i.m.Unlock()
 
@@ -709,8 +753,8 @@ func (i *installerImpl) Postinst(ctx context.Context, pkg string, caller string)
 	}
 }
 
-// Prerm runs the pre-rm script for a package.
-func (i *installerImpl) Prerm(ctx context.Context, pkg string, caller string, upgrade bool) error {
+// PreRemove runs the pre-rm command for a package.
+func (i *installerImpl) PreRemove(ctx context.Context, pkg string, caller string, upgrade bool) error {
 	i.m.Lock()
 	defer i.m.Unlock()
 
@@ -724,8 +768,85 @@ func (i *installerImpl) Prerm(ctx context.Context, pkg string, caller string, up
 		if caller == "deb" || caller == "rpm" {
 			installPath = "/opt/datadog-agent"
 		}
-		packages.PreRemoveAgent(ctx, installPath, caller, upgrade)
+		return datadogagent.PreRemove(ctx, installPath, caller, upgrade)
+	default:
 		return nil
+	}
+}
+
+// PreStartExperiment runs the pre-start-experiment command for a package.
+func (i *installerImpl) PreStartExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PreStartExperiment(ctx)
+	default:
+		return nil
+	}
+}
+
+// PostStartExperiment runs the post-start-experiment command for a package.
+func (i *installerImpl) PostStartExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PostStartExperiment(ctx)
+	default:
+		return nil
+	}
+}
+
+// PreStopExperiment runs the pre-stop-experiment command for a package.
+func (i *installerImpl) PreStopExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PreStopExperiment(ctx)
+	default:
+		return nil
+	}
+}
+
+// PostStopExperiment runs the post-stop-experiment command for a package.
+func (i *installerImpl) PostStopExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PostStopExperiment(ctx)
+	default:
+		return nil
+	}
+}
+
+// PrePromoteExperiment runs the pre-promote-experiment command for a package.
+func (i *installerImpl) PrePromoteExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PrePromoteExperiment(ctx)
+	default:
+		return nil
+	}
+}
+
+// PostPromoteExperiment runs the post-promote-experiment command for a package.
+func (i *installerImpl) PostPromoteExperiment(ctx context.Context, pkg string) error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PostPromoteExperiment(ctx)
 	default:
 		return nil
 	}
@@ -738,12 +859,33 @@ func (i *installerImpl) Close() error {
 	return i.close()
 }
 
-func (i *installerImpl) startExperiment(ctx context.Context, pkg string) error {
+func (i *installerImpl) preStartExperiment(ctx context.Context, pkg string) error {
+	switch pkg {
+	case packageDatadogAgent:
+		if runtime.GOOS == "windows" {
+			return nil
+		}
+		defer func() {
+			db, err := newDB()
+			if err != nil {
+				log.Errorf("could not recreate packages db: %v", err)
+			}
+			i.db = db
+		}()
+		i.db.Close()
+		installerExec := iexec.NewInstallerExec(i.env, filepath.Join(datadogagent.StablePath, "embedded", "bin", "installer"))
+		return installerExec.PreStartExperiment(ctx, pkg)
+	default:
+		return nil
+	}
+}
+
+func (i *installerImpl) postStartExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
 		// close so package can be updated as watchdog runs
 		i.db.Close()
-		return packages.StartAgentExperiment(ctx)
+		return datadogagent.PostStartExperiment(ctx)
 	case packageDatadogInstaller:
 		return packages.StartInstallerExperiment(ctx)
 	case packageAPMLibraryDotnet:
@@ -753,10 +895,10 @@ func (i *installerImpl) startExperiment(ctx context.Context, pkg string) error {
 	}
 }
 
-func (i *installerImpl) stopExperiment(ctx context.Context, pkg string) error {
+func (i *installerImpl) preStopExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return packages.StopAgentExperiment(ctx)
+		return datadogagent.PreStopExperiment(ctx)
 	case packageDatadogInstaller:
 		return packages.StopInstallerExperiment(ctx)
 	case packageAPMLibraryDotnet:
@@ -766,10 +908,52 @@ func (i *installerImpl) stopExperiment(ctx context.Context, pkg string) error {
 	}
 }
 
-func (i *installerImpl) promoteExperiment(ctx context.Context, pkg string) error {
+func (i *installerImpl) postStopExperiment(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return packages.PromoteAgentExperiment(ctx)
+		if runtime.GOOS == "windows" {
+			return nil
+		}
+		defer func() {
+			db, err := newDB()
+			if err != nil {
+				log.Errorf("could not recreate packages db: %v", err)
+			}
+			i.db = db
+		}()
+		i.db.Close()
+		installerExec := iexec.NewInstallerExec(i.env, filepath.Join(datadogagent.StablePath, "embedded", "bin", "installer"))
+		return installerExec.PostStopExperiment(ctx, pkg)
+	default:
+		return nil
+	}
+}
+
+func (i *installerImpl) prePromoteExperiment(ctx context.Context, pkg string) error {
+	switch pkg {
+	case packageDatadogAgent:
+		if runtime.GOOS == "windows" {
+			return nil
+		}
+		defer func() {
+			db, err := newDB()
+			if err != nil {
+				log.Errorf("could not recreate packages db: %v", err)
+			}
+			i.db = db
+		}()
+		i.db.Close()
+		installerExec := iexec.NewInstallerExec(i.env, filepath.Join(datadogagent.StablePath, "embedded", "bin", "installer"))
+		return installerExec.PrePromoteExperiment(ctx, pkg)
+	default:
+		return nil
+	}
+}
+
+func (i *installerImpl) postPromoteExperiment(ctx context.Context, pkg string) error {
+	switch pkg {
+	case packageDatadogAgent:
+		return datadogagent.PostPromoteExperiment(ctx)
 	case packageDatadogInstaller:
 		return packages.PromoteInstallerExperiment(ctx)
 	case packageAPMLibraryDotnet:
@@ -784,7 +968,7 @@ func (i *installerImpl) preparePackage(ctx context.Context, pkg string, _ []stri
 	case packageDatadogInstaller:
 		return packages.PrepareInstaller(ctx)
 	case packageDatadogAgent:
-		return packages.PrepareAgent(ctx)
+		return datadogagent.PreInstall(ctx)
 	default:
 		return nil
 	}
@@ -795,7 +979,7 @@ func (i *installerImpl) setupPackage(ctx context.Context, pkg string, args []str
 	case packageDatadogInstaller:
 		return packages.SetupInstaller(ctx)
 	case packageDatadogAgent:
-		return packages.SetupAgent(ctx, args)
+		return datadogagent.PostInstall(ctx, datadogagent.StablePath, "installer", args...)
 	case packageAPMInjector:
 		return packages.SetupAPMInjector(ctx)
 	case packageAPMLibraryDotnet:
@@ -808,7 +992,7 @@ func (i *installerImpl) setupPackage(ctx context.Context, pkg string, args []str
 func (i *installerImpl) removePackage(ctx context.Context, pkg string) error {
 	switch pkg {
 	case packageDatadogAgent:
-		return packages.RemoveAgent(ctx)
+		return datadogagent.PreRemove(ctx, datadogagent.StablePath, "installer", false)
 	case packageAPMInjector:
 		return packages.RemoveAPMInjector(ctx)
 	case packageDatadogInstaller:
